Split up classes to one-class-per-file to match style guide
