+++ Source/Editor/DualityEditor/Utility/DualityEditorUserData.cs
@@ -65,34 +65,4 @@ namespace Duality.Editor
 			set { this.pluginSettings = value; }
 		}
 	}
-
-	public class PluginSettings
-	{
-		private XElement oldStyleSettings = XElement.Parse("<Plugins></Plugins>");
-		private List<object> plugins = new List<object>();
-
-		[Obsolete("Use the model based api")]
-		public XElement OldStyleSettings
-		{
-			get { return this.oldStyleSettings; }
-			set { this.oldStyleSettings = value; }
-		}
-
-		public T GetSettings<T>() where T : class, new()
-		{
-			T setting = this.plugins.OfType<T>().FirstOrDefault();
-			if (setting == null)
-			{
-				setting = new T();
-				this.plugins.Add(setting);
-			}
-
-			return setting;
-		}
-		internal void Clear()
-		{
-			this.plugins.Clear();
-			this.oldStyleSettings = XElement.Parse("<Plugins></Plugins>");
-		}
-	}
 }
